requirements/install_requires.txt
django-geoads==0.0.2
django-floppyforms==1.1
django-rq-dashboard==0.2
django-pagination==1.0.7
django-userena==1.2.3
django-bootstrap-form==2.0.5
django-compressor==1.3
django-smart-extends==0.6
django-debug-toolbar==0.8.5
premailer==1.2.3
django-sorting<=0.1-1